chore: extract sticky resharding strategy to a new module
